refactor: move validation rules & types to separate files
